Test the CallViewModel output switcher directly
